UPDATE : count int->double에 따른 수정사항 반영
질문 관련 Entity의 count필드 타입 변경에 따른 타입 변경

+++ backend/src/main/java/com/wish/api/service/QuestionServiceImpl.java
@@ -30,7 +30,7 @@ public class QuestionServiceImpl implements QuestionService {
         
         if(ques.isPresent()){
             Question ques2 = ques.get();
-            int temp = ques2.getCount();
+            double temp = ques2.getCount();
             temp++;
             ques2.setCount(temp);
             questionRepository.save(ques2);
